separated exception types for one of and any of exceptions
